Updating tree tests
