+++ src/ENDFtk/tree/Tape/test/Tape.test.cpp
@@ -1,6 +1,7 @@
-#define CATCH_CONFIG_MAIN
+// include Catch2
+#include <catch2/catch_test_macros.hpp>
 
-#include "catch.hpp"
+// what we are testing
 #include "ENDFtk/tree/Tape.hpp"
 
 // other includes
